[WebPage] Serve esp.css inline, not by referring it
The first few times you may try to load esp.,css stored on the SPIFFS file system, you may not fetch the file at all.
This will render the web interface without CSS, so you have a hardly usable page.
This has to do with the fact the ESP can only handle a single transfer at once while the browser will try to load it almost immediately parallel to loading the rest of the page.
As soon as the browser has a copy of the CSS in its cache, this problem is over.

By loading the CSS inline, you will circumvent this issue.
However it will cause for a lot longer page loading times, so it is best to load the CSS from a define in the Custom.h