[ESPEasy-Now] Use status LED and allow sending without WiFi connection
The controllers now can also send if the ESPEasy-Now fallback option is enabled.
In normal use cases, one would like to use the ESPEasy-Now fallback option when there is no WiFi present. So it makes no sense to only allow to send when there is a WiFi connection.

The ESPEasyNowReceiver plugin now also adds the peers upon init of the plugin.